feat: Move cawg parsing for c2patool into own file
